add validation of user permission in controller and add a user permission menu

+++ application/controllers/announcement.php
@@ -16,42 +16,67 @@ class announcement extends CI_Controller {
 			redirect(base_url().'error_403');
 		}*/
 		$this->load->model('announcement_model');
+		$this->load->model('user_permision');
+		$this->load->library('session');
 	}
 
 	public function index() {
-		$data['msg'] = $this->session->flashdata('msg');
-		$b_url = base_url().'announcement/index';
-		$t_rows = $this->announcement_model->count();
-		$pageConfig = create_pagination_config( $b_url, $t_rows, 10, 3);
-		$this->pagination->initialize($pageConfig);
-		$page = ($this->uri->segment(3)) ? $this->uri->segment(3) : 0;
-		
-		$data['announcements'] = $this->announcement_model->fetch($pageConfig['per_page'], $page);
-		$data['links'] = $this->pagination->create_links();
 
-		$current_page =  floor(($this->uri->segment(3) / $pageConfig['per_page']) + 1);
-		$data['pagination_msg'] = create_pagination_msg($current_page, $pageConfig['per_page'], $t_rows);
+				$a = $this->user_permision->check_action_permision('announcement_view',$this->session->userdata('fcs_user_id'));
 
-		$this->load->view('_template/header', $data);
-        $this->load->view('announcement/index', $data);
-        $this->load->view('_template/footer', $data);
+
+			if($a['announcement_view'] == 0){
+
+				redirect(base_url().'error_550');
+		     }else{
+
+					$data['msg'] = $this->session->flashdata('msg');
+					$b_url = base_url().'announcement/index';
+					$t_rows = $this->announcement_model->count();
+					$pageConfig = create_pagination_config( $b_url, $t_rows, 10, 3);
+					$this->pagination->initialize($pageConfig);
+					$page = ($this->uri->segment(3)) ? $this->uri->segment(3) : 0;
+					
+					$data['announcements'] = $this->announcement_model->fetch($pageConfig['per_page'], $page);
+					$data['links'] = $this->pagination->create_links();
+
+					$current_page =  floor(($this->uri->segment(3) / $pageConfig['per_page']) + 1);
+					$data['pagination_msg'] = create_pagination_msg($current_page, $pageConfig['per_page'], $t_rows);
+
+					$this->load->view('_template/header', $data);
+			        $this->load->view('announcement/index', $data);
+			        $this->load->view('_template/footer', $data);
+
+			 }       
+
+
 	}
 	public function add(){
-		$this->load->helper(array('form'));
-		$this->load->library('form_validation');
-		$this->form_validation->set_rules('announce_title', 'supplier_name', 'required|callback_check_add_company_name');
 
-		if($this->form_validation->run() === FALSE) {
-			$data['action'] = 'add';
-			$this->load->view('_template/header', $data);
-			$this->load->view('announcement/add_edit', $data);
-			$this->load->view('_template/footer', $data);
-		}
-		else{
-			$this->announcement_model->add();
-			$this->session->set_flashdata('msg', 'New announcement successfully created');
-			redirect(base_url().'announcement');
-		}
+			$a = $this->user_permision->check_action_permision('announcement_add',$this->session->userdata('fcs_user_id'));
+
+
+			if($a['announcement_add'] == 0){
+
+				redirect(base_url().'error_550');
+		     }else{		
+				$this->load->helper(array('form'));
+				$this->load->library('form_validation');
+				$this->form_validation->set_rules('announce_title', 'supplier_name', 'required|callback_check_add_company_name');
+
+				if($this->form_validation->run() === FALSE) {
+					$data['action'] = 'add';
+					$this->load->view('_template/header', $data);
+					$this->load->view('announcement/add_edit', $data);
+					$this->load->view('_template/footer', $data);
+				}
+				else{
+					$this->announcement_model->add();
+					$this->session->set_flashdata('msg', 'New announcement successfully created');
+					redirect(base_url().'announcement');
+				}
+
+			}	
 	}
 
 	public function check_add_company_name(){
@@ -66,26 +91,40 @@ class announcement extends CI_Controller {
 	}
 
 	public function edit($id){
-		$data['announcement'] = $this->announcement_model->get($id);
-		if (empty($data['announcement'])) {
-			show_404();
-		}
-		$this->load->helper('form');
-		$this->load->library('form_validation');
 
-		$this->form_validation->set_rules('announce_title', 'category', 'required|callback_check_edit_company_name['.$id.']');
+			$a = $this->user_permision->check_action_permision('announcement_edit',$this->session->userdata('fcs_user_id'));
 
-		if($this->form_validation->run() === FALSE) {
-			$data['action'] = 'edit';
-			$this->load->view('_template/header', $data);
-			$this->load->view('announcement/add_edit', $data);
-			$this->load->view('_template/footer', $data);	
-		}
-		else {
-			$this->announcement_model->update($id);
-			$this->session->set_flashdata('msg', 'announcement successfully updated');
-			redirect(base_url().'announcement');
-		}
+
+			if($a['announcement_edit'] == 1){
+
+				redirect(base_url().'error_550');
+		     }else{	
+
+
+				$data['announcement'] = $this->announcement_model->get($id);
+				if (empty($data['announcement'])) {
+					show_404();
+				}
+				$this->load->helper('form');
+				$this->load->library('form_validation');
+
+				$this->form_validation->set_rules('announce_title', 'category', 'required|callback_check_edit_company_name['.$id.']');
+
+				if($this->form_validation->run() === FALSE) {
+					$data['action'] = 'edit';
+					$this->load->view('_template/header', $data);
+					$this->load->view('announcement/add_edit', $data);
+					$this->load->view('_template/footer', $data);	
+				}
+				else {
+					$this->announcement_model->update($id);
+					$this->session->set_flashdata('msg', 'announcement successfully updated');
+					redirect(base_url().'announcement');
+				}
+
+			}	
+
+
 	}
 	public function check_edit_company_name($str, $id){
 		$company_name = $this->input->post('announce_title');
@@ -98,11 +137,21 @@ class announcement extends CI_Controller {
 	}
 
 	public function delete($id="") {
-		if ($id=="") {
-			show_404();
-		}
-		$this->announcement_model->delete($id);	
-		$this->session->set_flashdata('msg', 'announcement successfully deleted');
-		redirect(base_url().'announcement');
+
+			$a = $this->user_permision->check_action_permision('announcement_del',$this->session->userdata('fcs_user_id'));
+
+
+			if($a['announcement_del'] == 1){
+
+				redirect(base_url().'error_550');
+		     }else{	
+
+				if ($id=="") {
+					show_404();
+				}
+				$this->announcement_model->delete($id);	
+				$this->session->set_flashdata('msg', 'announcement successfully deleted');
+				redirect(base_url().'announcement');
+			}		
 	}
 }
