+++ application/controllers/daily_sales.php
@@ -14,26 +14,43 @@ class daily_sales extends CI_Controller {
 		//{
 		//	redirect(base_url().'error_403');
 		//}
+		$this->load->model('insurance_package_model');
 		$this->load->model('daily_sales_model');
+		$this->load->model('user_permision');
+		$this->load->library('session');
 	}
 
 	public function index() {
-		$data['msg'] = $this->session->flashdata('msg');
-		$b_url = base_url().'daily_sales/index';
-		$t_rows = $this->daily_sales_model->count();
-		$pageConfig = create_pagination_config( $b_url, $t_rows, 10, 3);
-		$this->pagination->initialize($pageConfig);
-		$page = ($this->uri->segment(3)) ? $this->uri->segment(3) : 0;
-		
-		$data['daily_salesx'] = $this->daily_sales_model->fetch($pageConfig['per_page'], $page);
-		$data['links'] = $this->pagination->create_links();
 
-		$current_page =  floor(($this->uri->segment(3) / $pageConfig['per_page']) + 1);
-		$data['pagination_msg'] = create_pagination_msg($current_page, $pageConfig['per_page'], $t_rows);
 
-		$this->load->view('_template/header', $data);
-        $this->load->view('daily_sales/index', $data);
-        $this->load->view('_template/footer', $data);
+			$a = $this->user_permision->check_action_permision('dailysales_view',$this->session->userdata('fcs_user_id'));
+
+
+			if($a['dailysales_view'] == 0){
+
+				redirect(base_url().'error_550');
+		     }else{
+
+				$data['msg'] = $this->session->flashdata('msg');
+				$b_url = base_url().'daily_sales/index';
+				$t_rows = $this->daily_sales_model->count();
+				$pageConfig = create_pagination_config( $b_url, $t_rows, 10, 3);
+				$this->pagination->initialize($pageConfig);
+				$page = ($this->uri->segment(3)) ? $this->uri->segment(3) : 0;
+				
+				$data['daily_salesx'] = $this->daily_sales_model->fetch($pageConfig['per_page'], $page);
+				$data['links'] = $this->pagination->create_links();
+
+				$current_page =  floor(($this->uri->segment(3) / $pageConfig['per_page']) + 1);
+				$data['pagination_msg'] = create_pagination_msg($current_page, $pageConfig['per_page'], $t_rows);
+
+				$this->load->view('_template/header', $data);
+		        $this->load->view('daily_sales/index', $data);
+		        $this->load->view('_template/footer', $data);
+
+		    }    
+
+
 	}
 	public function add(){
 		$this->load->helper(array('form'));
@@ -65,26 +82,38 @@ class daily_sales extends CI_Controller {
 	}
 
 	public function edit($id){
-		$data['insurance_package'] = $this->insurance_package_model->get($id);
-		if (empty($data['insurance_package'])) {
-			show_404();
-		}
-		$this->load->helper('form');
-		$this->load->library('form_validation');
 
-		$this->form_validation->set_rules('insurance_name', 'insurance_name', 'required|callback_check_edit_company_name['.$id.']');
+			$a = $this->user_permision->check_action_permision('dailysales_edit',$this->session->userdata('fcs_user_id'));
 
-		if($this->form_validation->run() === FALSE) {
-			$data['action'] = 'edit';
-			$this->load->view('_template/header', $data);
-			$this->load->view('insurance_package/add_edit', $data);
-			$this->load->view('_template/footer', $data);	
-		}
-		else {
-			$this->insurance_package_model->update($id);
-			$this->session->set_flashdata('msg', 'Insurance successfully updated');
-			redirect(base_url().'insurance_package');
-		}
+
+			if($a['dailysales_edit'] == 0){
+
+				redirect(base_url().'error_550');
+		     }else{
+
+					$data['insurance_package'] = $this->insurance_package_model->get($id);
+					if (empty($data['insurance_package'])) {
+						show_404();
+					}
+					$this->load->helper('form');
+					$this->load->library('form_validation');
+
+					$this->form_validation->set_rules('insurance_name', 'insurance_name', 'required|callback_check_edit_company_name['.$id.']');
+
+					if($this->form_validation->run() === FALSE) {
+						$data['action'] = 'edit';
+						$this->load->view('_template/header', $data);
+						$this->load->view('insurance_package/add_edit', $data);
+						$this->load->view('_template/footer', $data);	
+					}
+					else {
+						$this->insurance_package_model->update($id);
+						$this->session->set_flashdata('msg', 'Insurance successfully updated');
+						redirect(base_url().'insurance_package');
+					}
+
+			}		
+
 	}
 	public function check_edit_company_name($str, $id){
 		$insurance_name = $this->input->post('insurance_name');
@@ -97,11 +126,22 @@ class daily_sales extends CI_Controller {
 	}
 
 	public function delete($id="") {
-		if ($id=="") {
-			show_404();
-		}
-		$this->insurance_package_model->delete($id);	
-		$this->session->set_flashdata('msg', 'Insurance successfully deleted');
-		redirect(base_url().'insurance_package');
+
+			$a = $this->user_permision->check_action_permision('dailysales_del',$this->session->userdata('fcs_user_id'));
+
+
+			if($a['dailysales_del'] == 1){
+
+				redirect(base_url().'error_550');
+		     }else{
+
+				if ($id=="") {
+					show_404();
+				}
+				$this->insurance_package_model->delete($id);	
+				$this->session->set_flashdata('msg', 'Insurance successfully deleted');
+				redirect(base_url().'insurance_package');
+
+			}	
 	}
 }
